requirements.txt
ddddocr
requests
Pillow
cairosvg
